Moved EMQ logic into BufferManagerImpl
