+++ dev/com.ibm.ws.logging/src/com/ibm/ws/logging/internal/impl/BaseTraceService.java
@@ -26,6 +26,7 @@ import com.ibm.websphere.logging.WsLevel;
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.TruncatableThrowable;
+import com.ibm.ws.collector.manager.buffer.BufferManagerImpl;
 import com.ibm.ws.collector.manager.buffer.SimpleRotatingSoftQueue;
 import com.ibm.ws.kernel.boot.logging.LoggerHandlerManager;
 import com.ibm.ws.logging.RoutedMessage;
@@ -37,7 +38,6 @@ import com.ibm.ws.logging.internal.TraceSpecification;
 import com.ibm.ws.logging.internal.WsLogRecord;
 import com.ibm.ws.logging.utils.FileLogHolder;
 import com.ibm.ws.logging.utils.RecursionCounter;
-import com.ibm.wsspi.collector.manager.BufferManager;
 import com.ibm.wsspi.logging.LogHandler;
 import com.ibm.wsspi.logging.MessageRouter;
 import com.ibm.wsspi.logprovider.LogProviderConfig;
@@ -215,7 +215,7 @@ public class BaseTraceService implements TrService {
         // LogManager, which might print errors due to misconfiguration.
         captureSystemStreams();
         //Remove EMQ from BufferManager after a certain amount of time has passed
-        BufferManager.removeEMQByTimer();
+        BufferManagerImpl.removeEMQByTimer();
     }
 
     protected void registerLoggerHandlerSingleton() {
