+++ dev/com.ibm.ws.logging/src/com/ibm/ws/logging/source/LogSource.java
@@ -18,6 +18,7 @@ import java.util.regex.Pattern;
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Trivial;
+import com.ibm.ws.collector.manager.buffer.BufferManagerImpl;
 import com.ibm.ws.logging.RoutedMessage;
 import com.ibm.ws.logging.WsLogHandler;
 import com.ibm.ws.logging.internal.WsLogRecord;
@@ -100,8 +101,8 @@ public class LogSource implements Source, WsLogHandler {
             LogRecord logRecord = routedMessage.getLogRecord();
             if (logRecord != null && bufferMgr != null) {
                 MessageLogData parsedMessage = parse(routedMessage, logRecord);
-                if (!BufferManager.getEMQRemovedFlag() && parsedMessage.getMessage().startsWith("CWWKF0011I")) {
-                    BufferManager.removeEMQTrigger();
+                if (!BufferManagerImpl.getEMQRemovedFlag() && parsedMessage.getMessage().startsWith("CWWKF0011I")) {
+                    BufferManagerImpl.removeEMQTrigger();
                 }
                 bufferMgr.add(parsedMessage);
             }
